requirements.txt
django==4.2.7
google-generativeai==0.8.5
python-dotenv==1.0.0
requests==2.31.0
